Refresh user's repos one-by-one (#29)
dockstore/dockstore#3522

This is mainly an upgrade to 1.9, the need of which only became apparent after I started the actual feature.

Changes for this feature
------------------------

The summary printed out at the end of a refresh had a column named
"ON DOCKSTORE?" which I renamed to "PUBLISHED". They're all on
 Dockstore, IMHO.

Failures to refresh individual repos do not cause the refresh to fail overall; an error message is printed out, but the CLI continues to refresh repos.

Created dockstore/dockstore#3571 to handle discovery of repos the user has been added to.

1.9 upgrade changes
---------------------

Had to update a bunch of version dependencies for build to work against 1.9.
Move to postgres 11
Had to update a lot of SQL that was using the workflow_workflowversions table that no longer exists.
* I don't understand decrypt.sh used to work, but updated it to match profile names. It was matching confidential-tests but we have a profile named non-confidential-tests. And Singularity doesn't have "confidential" in the profile name.
    Since refresh org endpoint doesn't exist, copied refreshByOrganizationReplacement from Denis' web service PR and using that.
    Lots of code needed to change from Registry.QUAY_IO.toString to Registry.QUAY_IO.getDockerPath() -- a web service change necessitated that.
    Created dockstore/dockstore#3569 and commented out a test until it's fixed.
    Created dockstore/dockstore#3563 -- workaround for now is to make two API calls.

Other
------

Occasionally ran out of memory on Circle CI; taking advantage of our new Performance plan!